+++ CHANGELOG.md
@@ -3,9 +3,13 @@ cloudstack_wrapper CHANGELOG
 
 This file is used to list changes made in each version of the cloudstack_wrapper cookbook.
 
+0.2.7
+-----
+- ccontini - Handle vhost definition as an environment attribute.
+
 0.2.6
 -----
-- ccontini - Add eventlog config
+- ccontini - Add eventlog config.
 
 0.2.5
 -----